Restructured build
