+++ src/pycarl/core/bound_type.cpp
