+++ src/pycarl/core/bound_type.h
