+++ src/pycarl/core/variable.cpp
