+++ src/pycarl/core/variable.h
